Reorganize group formula javascript files to its own feature folder
