+++ java/code/src/com/suse/manager/webui/templates/groups/formula.jade
@@ -20,4 +20,4 @@ script(type='text/javascript').
     var groupId = "#{groupId}";
     var formulaId = #{formula_id};
 
-script(src='/javascript/manager/group-formula.bundle.js?cb=#{webVersion}', type='text/javascript')
+script(src='/javascript/manager/groups/formula/group-formula.bundle.js?cb=#{webVersion}', type='text/javascript')
