+++ java/code/src/com/suse/manager/webui/templates/groups/formulas.jade
@@ -19,4 +19,4 @@ script(type='text/javascript').
     var csrfToken = "#{csrf_token}";
     var groupId = "#{groupId}";
 
-script(src='/javascript/manager/group-formula-selection.bundle.js?cb=#{webVersion}', type='text/javascript')
+script(src='/javascript/manager/groups/formula/group-formula-selection.bundle.js?cb=#{webVersion}', type='text/javascript')
