+++ web/html/src/manager/groups/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  entries: [
+    'formula/group-formula.js',
+    'formula/group-formula-selection.js',
+  ]
+};
